# Milestone 4/backend/swagger.yaml
openapi: 3.1.0
info:
  title: EduAssist API
  description: AI-Powered Teaching Assistant Platform - Knowledge Assistant Feature
  version: 1.0.0
paths:
...
  /study-guides/study-guides/:
    post:
      tags:
        - Study Guides
        - Study Guide Generator
      summary: Create Study Guide
      description: |-
        Create a study guide from YouTube video

        Steps:
        1. Validate YouTube URL and get video ID
        2. Get video metadata
        3. Create study guide record (PROCESSING status)
        4. Generate content in background (async)
        5. Return study guide ID immediately

        The client should poll /study-guides/{id} to check status
      operationId: create_study_guide_study_guides_study_guides__post
      security:
        - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudyGuideCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyGuideResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - Study Guides
        - Study Guide Generator
      summary: List Study Guides
      description: List study guides
      operationId: list_study_guides_study_guides_study_guides__get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: course_id
          in: query
          required: false
          schema:
            type: integer
            title: Course Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyGuideListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /assessments/assessments/:
    post:
      tags:
        - Assessments
        - Assessment Generator
      summary: Create Assessment
      description: |-
        Create AI-generated assessment

        Steps:
        1. Create assessment record (GENERATING status)
        2. Generate questions in background
        3. Return assessment ID immediately

        Client should poll /assessments/{id} to check status
      operationId: create_assessment_assessments_assessments__post
      security:
        - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssessmentCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - Assessments
        - Assessment Generator
      summary: List Assessments
      description: List assessments
      operationId: list_assessments_assessments_assessments__get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: course_id
          in: query
          required: false
          schema:
            type: integer
            title: Course Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /slide-decks/slide-decks/:
    post:
      tags:
        - Slide Decks
        - Slide Decks
      summary: Start Slide Deck
      description: Generate slide deck from uploaded docs. Returns deck_id (status PENDING/COMPLETE)
      operationId: start_slide_deck_slide_decks_slide_decks__post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SlideDeckCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Start Slide Deck Slide Decks Slide Decks  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - Slide Decks
        - Slide Decks
      summary: List Decks
      operationId: list_decks_slide_decks_slide_decks__get
      parameters:
        - name: status
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SlideDeckOut'
                title: Response List Decks Slide Decks Slide Decks  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /feedback/:
    post:
      tags:
        - Feedback
      summary: Submit Feedback
      operationId: submit_feedback_feedback__post
      parameters:
        - name: user_id
          in: query
          required: false
          schema:
            type: integer
            default: 1
            title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - Feedback
      summary: Get Feedback
      operationId: get_feedback_feedback__get
      parameters:
        - name: target_id
          in: query
          required: true
          schema:
            type: integer
            title: Target Id
        - name: target_type
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/FeedbackTargetType'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeedbackResponse'
                title: Response Get Feedback Feedback  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/register:
    post:
      tags:
        - Auth
      summary: Register
      operationId: register_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /token:
    post:
      summary: Login
      description: Login and get access token
      operationId: login_token_post
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Body_login_token_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /register:
    post:
      summary: Register
      description: Register a new user
      operationId: register_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /courses:
    get:
      summary: List Courses
      description: List all courses
      operationId: list_courses_courses_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CourseResponse'
                type: array
                title: Response List Courses Courses Get
    post:
      summary: Create Course
      description: Create a new course
      operationId: create_course_courses_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AssessmentCreate:
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 1
          title: Title
        description:
          anyOf:
            - type: string
            - type: 'null'
          title: Description
        course_id:
          type: integer
          title: Course Id
        custom_prompt:
          type: string
          minLength: 10
          title: Custom Prompt
          description: Instructions for question generation
        difficulty_level:
          $ref: '#/components/schemas/DifficultyLevel'
          default: MEDIUM
        question_types:
          items:
            $ref: '#/components/schemas/QuestionType'
          type: array
          minItems: 1
          title: Question Types
        total_questions:
          type: integer
          maximum: 100
          minimum: 1
          title: Total Questions
        reference_document_ids:
          items:
            type: integer
          type: array
          title: Reference Document Ids
          description: Document IDs to use as reference
        duration_minutes:
          anyOf:
            - type: integer
              maximum: 600
              minimum: 5
            - type: 'null'
          title: Duration Minutes
      type: object
      required:
        - title
        - course_id
        - custom_prompt
        - question_types
        - total_questions
      title: AssessmentCreate
      description: Request to create assessment
    AssessmentListResponse:
      properties:
        assessments:
          items:
            $ref: '#/components/schemas/AssessmentResponse'
          type: array
          title: Assessments
        total:
          type: integer
          title: Total
      type: object
      required:
        - assessments
        - total
      title: AssessmentListResponse
      description: List of assessments
...
    AssessmentResponse:
      properties:
        id:
          type: integer
          title: Id
        title:
          type: string
          title: Title
        description:
          anyOf:
            - type: string
            - type: 'null'
          title: Description
        custom_prompt:
          type: string
          title: Custom Prompt
        difficulty_level:
          $ref: '#/components/schemas/DifficultyLevel'
        question_types:
          items:
            type: string
          type: array
          title: Question Types
        total_questions:
          type: integer
          title: Total Questions
        status:
          $ref: '#/components/schemas/AssessmentStatus'
        total_marks:
          anyOf:
            - type: integer
            - type: 'null'
          title: Total Marks
        duration_minutes:
          anyOf:
            - type: integer
            - type: 'null'
          title: Duration Minutes
        course_id:
          type: integer
          title: Course Id
        created_by:
          type: integer
          title: Created By
        created_at:
          type: string
          format: date-time
          title: Created At
        questions:
          items:
            $ref: '#/components/schemas/QuestionResponse'
          type: array
          title: Questions
          default: []
      type: object
      required:
        - id
        - title
        - custom_prompt
        - difficulty_level
        - question_types
        - total_questions
        - status
        - course_id
        - created_by
        - created_at
      title: AssessmentResponse
      description: Assessment response
    AssessmentStatus:
      type: string
      enum:
        - GENERATING
        - COMPLETED
        - FAILED
      title: AssessmentStatus
...
    Body_login_token_post:
      properties:
        grant_type:
          anyOf:
            - type: string
              pattern: password
            - type: 'null'
          title: Grant Type
        username:
          type: string
          title: Username
        password:
          type: string
          title: Password
        scope:
          type: string
          title: Scope
          default: ''
        client_id:
          anyOf:
            - type: string
            - type: 'null'
          title: Client Id
        client_secret:
          anyOf:
            - type: string
            - type: 'null'
          title: Client Secret
      type: object
      required:
        - username
        - password
      title: Body_login_token_post
...
    CourseCreate:
      properties:
        code:
          type: string
          maxLength: 20
          minLength: 1
          title: Code
        name:
          type: string
          maxLength: 255
          minLength: 1
          title: Name
        description:
          anyOf:
            - type: string
            - type: 'null'
          title: Description
      type: object
      required:
        - code
        - name
      title: CourseCreate
    CourseResponse:
      properties:
        id:
          type: integer
          title: Id
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        description:
          anyOf:
            - type: string
            - type: 'null'
          title: Description
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
        - id
        - code
        - name
        - created_at
      title: CourseResponse
    DifficultyLevel:
      type: string
      enum:
        - EASY
        - MEDIUM
        - HARD
      title: DifficultyLevel
...
    FeedbackCreate:
      properties:
        target_id:
          type: integer
          title: Target Id
        target_type:
          $ref: '#/components/schemas/FeedbackTargetType'
        rating:
          type: number
          maximum: 5
          minimum: 1
          title: Rating
        aspect_ratings:
          anyOf:
            - additionalProperties:
                type: number
              type: object
            - type: 'null'
          title: Aspect Ratings
        comment:
          anyOf:
            - type: string
            - type: 'null'
          title: Comment
      type: object
      required:
        - target_id
        - target_type
        - rating
      title: FeedbackCreate
    FeedbackResponse:
      properties:
        target_id:
          type: integer
          title: Target Id
        target_type:
          $ref: '#/components/schemas/FeedbackTargetType'
        rating:
          type: number
          maximum: 5
          minimum: 1
          title: Rating
        aspect_ratings:
          anyOf:
            - additionalProperties:
                type: number
              type: object
            - type: 'null'
          title: Aspect Ratings
        comment:
          anyOf:
            - type: string
            - type: 'null'
          title: Comment
        id:
          type: integer
          title: Id
        user_id:
          type: integer
          title: User Id
        created_at:
          type: string
          title: Created At
      type: object
      required:
        - target_id
        - target_type
        - rating
        - id
        - user_id
        - created_at
      title: FeedbackResponse
    FeedbackTargetType:
      type: string
      enum:
        - assessment
        - study_guide
        - slide_deck
      title: FeedbackTargetType
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Priority:
      type: string
      enum:
        - high
        - medium
        - low
      title: Priority
    PrioritySegment:
      properties:
        start:
          type: integer
          minimum: 0
          title: Start
          description: Start time in seconds
        end:
          type: integer
          exclusiveMinimum: 0
          title: End
          description: End time in seconds
        priority:
          $ref: '#/components/schemas/Priority'
          default: medium
      type: object
      required:
        - start
        - end
      title: PrioritySegment
      description: User-selected important segment
...
    QuestionResponse:
      properties:
        id:
          type: integer
          title: Id
        assessment_id:
          type: integer
          title: Assessment Id
        question_type:
          $ref: '#/components/schemas/QuestionType'
        question_text:
          type: string
          title: Question Text
        options:
          anyOf:
            - items:
                type: string
              type: array
            - type: 'null'
          title: Options
        correct_answer:
          title: Correct Answer
        explanation:
          anyOf:
            - type: string
            - type: 'null'
          title: Explanation
        marks:
          type: integer
          title: Marks
        difficulty:
          $ref: '#/components/schemas/DifficultyLevel'
        order:
          type: integer
          title: Order
      type: object
      required:
        - id
        - assessment_id
        - question_type
        - question_text
        - correct_answer
        - marks
        - difficulty
        - order
      title: QuestionResponse
      description: Question response
    QuestionType:
      type: string
      enum:
        - MCQ
        - MSQ
        - NAT
        - SHORT_ANSWER
        - TRUE_FALSE
      title: QuestionType
...
    SlideDeckControls:
      properties:
        num_slides:
          type: integer
          maximum: 30
          minimum: 3
          title: Num Slides
          default: 10
        desired_depth:
          type: string
          title: Desired Depth
          description: summary, normal, in-depth
          default: summary
      type: object
      title: SlideDeckControls
    SlideDeckCreate:
      properties:
        title:
          type: string
          title: Title
        controls:
          $ref: '#/components/schemas/SlideDeckControls'
        reference_document_ids:
          anyOf:
            - items:
                type: integer
              type: array
            - type: 'null'
          title: Reference Document Ids
          default: []
      type: object
      required:
        - title
        - controls
      title: SlideDeckCreate
    SlideDeckOut:
      properties:
        id:
          type: integer
          title: Id
        title:
          type: string
          title: Title
        status:
          $ref: '#/components/schemas/SlideDeckStatus'
        controls:
          $ref: '#/components/schemas/SlideDeckControls'
        slides:
          items:
            $ref: '#/components/schemas/SlideOut'
          type: array
          title: Slides
          default: []
        created_at:
          type: string
          title: Created At
      type: object
      required:
        - id
        - title
        - status
        - controls
        - created_at
      title: SlideDeckOut
    SlideDeckStatus:
      type: string
      enum:
        - PENDING
        - GENERATING
        - COMPLETE
        - FAILED
      title: SlideDeckStatus
    SlideOut:
      properties:
        title:
          type: string
          title: Title
        content_md:
          type: string
          title: Content Md
        notes_md:
          anyOf:
            - type: string
            - type: 'null'
          title: Notes Md
        image_url:
          anyOf:
            - type: string
            - type: 'null'
          title: Image Url
        position:
          type: integer
          title: Position
        id:
          type: integer
          title: Id
      type: object
      required:
        - title
        - content_md
        - position
        - id
      title: SlideOut
    StudyGuideCreate:
      properties:
        youtube_url:
          type: string
          minLength: 1
          title: Youtube Url
        course_id:
          type: integer
          title: Course Id
        title:
          type: string
          maxLength: 255
          minLength: 1
          title: Title
        priority_segments:
          items:
            $ref: '#/components/schemas/PrioritySegment'
          type: array
          minItems: 1
          title: Priority Segments
      type: object
      required:
        - youtube_url
        - course_id
        - title
        - priority_segments
      title: StudyGuideCreate
      description: Request to create study guide from YouTube
    StudyGuideListResponse:
      properties:
        guides:
          items:
            $ref: '#/components/schemas/StudyGuideResponse'
          type: array
          title: Guides
        total:
          type: integer
          title: Total
      type: object
      required:
        - guides
        - total
      title: StudyGuideListResponse
      description: List of study guides
    StudyGuideResponse:
      properties:
        id:
          type: integer
          title: Id
        youtube_url:
          type: string
          title: Youtube Url
        video_id:
          type: string
          title: Video Id
        video_title:
          anyOf:
            - type: string
            - type: 'null'
          title: Video Title
        video_duration:
          anyOf:
            - type: integer
            - type: 'null'
          title: Video Duration
        title:
          type: string
          title: Title
        content:
          anyOf:
            - type: string
            - type: 'null'
          title: Content
        key_topics:
          anyOf:
            - items:
                type: string
              type: array
            - type: 'null'
          title: Key Topics
        status:
          $ref: '#/components/schemas/StudyGuideStatus'
        course_id:
          type: integer
          title: Course Id
        created_by:
          type: integer
          title: Created By
        created_at:
          type: string
          format: date-time
          title: Created At
        segments:
          items:
            $ref: '#/components/schemas/VideoSegmentResponse'
          type: array
          title: Segments
          default: []
      type: object
      required:
        - id
        - youtube_url
        - video_id
        - title
        - status
        - course_id
        - created_by
        - created_at
      title: StudyGuideResponse
      description: Study guide response
    StudyGuideStatus:
      type: string
      enum:
        - PROCESSING
        - COMPLETED
        - FAILED
      title: StudyGuideStatus
    Token:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
      type: object
      required:
        - access_token
        - token_type
      title: Token
    UserCreate:
      properties:
        email:
          type: string
          format: email
          title: Email
        username:
          type: string
          title: Username
        full_name:
          anyOf:
            - type: string
            - type: 'null'
          title: Full Name
        password:
          type: string
          title: Password
      type: object
      required:
        - email
        - username
        - password
      title: UserCreate
    UserResponse:
      properties:
        email:
          type: string
          format: email
          title: Email
        username:
          type: string
          title: Username
        full_name:
          anyOf:
            - type: string
            - type: 'null'
          title: Full Name
        id:
          type: integer
          title: Id
        is_active:
          type: boolean
          title: Is Active
        is_student:
          type: boolean
          title: Is Student
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
        - email
        - username
        - id
        - is_active
        - is_student
        - created_at
      title: UserResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
    VideoSegmentResponse:
      properties:
        id:
          type: integer
          title: Id
        start_time:
          type: integer
          title: Start Time
        end_time:
          type: integer
          title: End Time
        transcript:
          type: string
          title: Transcript
        priority:
          type: string
          title: Priority
        summary:
          anyOf:
            - type: string
            - type: 'null'
          title: Summary
        key_points:
          anyOf:
            - items:
                type: string
              type: array
            - type: 'null'
          title: Key Points
      type: object
      required:
        - id
        - start_time
        - end_time
        - transcript
        - priority
      title: VideoSegmentResponse
      description: Video segment with transcript
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: token